+++ soffa-foundation-core/src/main/java/io/soffa/foundation/core/actions/ActionRegistry.java
@@ -1,9 +1,0 @@
-package io.soffa.foundation.core.actions;
-
-import io.soffa.foundation.core.model.Validatable;
-
-public interface ActionRegistry {
-
-    <I extends Validatable, O> Action<I, O> lookup(Class<? extends Action<I, O>> action);
-
-}