+++ soffa-foundation-springboot/src/main/java/ext/springboot/RabbitMQConnectionFactory.java
@@ -1,19 +1,17 @@
 package ext.springboot;
 
 import io.soffa.foundation.spring.config.amqp.model.RabbitMQProperties;
-import org.springframework.boot.autoconfigure.AutoConfigureOrder;
 import org.springframework.boot.autoconfigure.amqp.RabbitAutoConfiguration;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Import;
-import org.springframework.core.Ordered;
 
 @Configuration
 @ConditionalOnProperty(value = "app.amqp.enabled", havingValue = "true")
 @Import(RabbitAutoConfiguration.class)
-@AutoConfigureOrder(Ordered.LOWEST_PRECEDENCE)
+//@AutoConfigureOrder(Ordered.LOWEST_PRECEDENCE)
 public class RabbitMQConnectionFactory {
 
     @Bean