+++ soffa-foundation-springboot/src/main/java/io/soffa/foundation/spring/DefaultActionDispatcher.java
@@ -1,10 +1,13 @@
 package io.soffa.foundation.spring;
 
+import com.google.common.collect.ImmutableMap;
 import io.soffa.foundation.context.RequestContextHolder;
 import io.soffa.foundation.core.RequestContext;
 import io.soffa.foundation.core.actions.Action;
 import io.soffa.foundation.core.actions.Action0;
 import io.soffa.foundation.core.actions.ActionDispatcher;
+import io.soffa.foundation.core.metrics.CoreMetrics;
+import io.soffa.foundation.core.metrics.MetricsRegistry;
 import io.soffa.foundation.core.model.Validatable;
 import io.soffa.foundation.exceptions.TechnicalException;
 import lombok.AllArgsConstructor;
@@ -12,13 +15,15 @@ import org.springframework.stereotype.Component;
 
 import javax.transaction.Transactional;
 
+import static io.soffa.foundation.core.metrics.CoreMetrics.ACTION_HANDLE;
+
 @Component
 @AllArgsConstructor
 public class DefaultActionDispatcher implements ActionDispatcher {
 
     private final ActionsMapping mapping;
+    private final MetricsRegistry metricsRegistry;
 
-    @SuppressWarnings("unchecked")
     @Override
     public <I, O> O dispatch(Class<? extends Action<I, O>> actionClass, I request) {
         RequestContext context = RequestContextHolder.require();
@@ -34,9 +39,12 @@ public class DefaultActionDispatcher implements ActionDispatcher {
         for (Action<?, ?> act : mapping.getRegistry()) {
             if (actionClass.isAssignableFrom(act.getClass())) {
                 Action<I, O> impl = (Action<I, O>) act;
-                return impl.handle(request, context);
+                return metricsRegistry.track(ACTION_HANDLE, ImmutableMap.of(
+                    "action", actionClass.getName()
+                ), () -> impl.handle(request, context));
             }
         }
+        metricsRegistry.increment(CoreMetrics.INVALID_ACTION);
         throw new TechnicalException("Unable to find implementation for action: %s", actionClass.getName());
     }
 
@@ -53,9 +61,12 @@ public class DefaultActionDispatcher implements ActionDispatcher {
         for (Action0<?> act : mapping.getRegistry0()) {
             if (actionClass.isAssignableFrom(act.getClass())) {
                 Action0<O> impl = (Action0<O>) act;
-                return impl.handle(context);
+                return metricsRegistry.track(ACTION_HANDLE, ImmutableMap.of(
+                    "action", actionClass.getName()
+                ), () -> impl.handle(context));
             }
         }
+        metricsRegistry.increment(CoreMetrics.INVALID_ACTION);
         throw new TechnicalException("Unable to find implementation for action: %s", actionClass.getName());
     }