+++ soffa-foundation-springboot/src/main/java/io/soffa/foundation/spring/aop/MessageDispatcherAspect.java
@@ -1,4 +1,4 @@
-package io.soffa.foundation.spring.config;
+package io.soffa.foundation.spring.aop;
 
 import io.soffa.foundation.annotations.DispatchMessage;
 import io.soffa.foundation.commons.Logger;
@@ -8,7 +8,6 @@ import lombok.SneakyThrows;
 import org.aspectj.lang.ProceedingJoinPoint;
 import org.aspectj.lang.annotation.Around;
 import org.aspectj.lang.annotation.Aspect;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
 @Aspect
@@ -19,7 +18,7 @@ public class MessageDispatcherAspect {
     private final MessageDispatcher messageDispatcher;
     public static final Logger LOG = Logger.get(MessageDispatcherAspect.class);
 
-    public MessageDispatcherAspect(@Autowired(required = false) MessageDispatcher messageDispatcher) {
+    public MessageDispatcherAspect(MessageDispatcher messageDispatcher) {
         this.messageDispatcher = messageDispatcher;
     }