+++ soffa-foundation-test-app/src/main/java/com/company/app/gateways/API.java
@@ -5,7 +5,7 @@ import com.company.app.core.PingAction;
 import com.company.app.core.PingResponse;
 import io.soffa.foundation.core.ApiHeaders;
 import io.soffa.foundation.core.RequestContext;
-import io.soffa.foundation.core.annotations.ActionBind;
+import io.soffa.foundation.core.annotations.BindAction;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.tags.Tag;
@@ -25,7 +25,7 @@ public interface API {
         parameters = {@Parameter(ref = ApiHeaders.TENANT_ID)}
     )
     @Path("/ping")
-    @ActionBind(PingAction.class)
+    @BindAction(PingAction.class)
     PingResponse ping(@Parameter(hidden = true) RequestContext context);
 
     @Operation(
@@ -35,7 +35,7 @@ public interface API {
         parameters = {@Parameter(ref = ApiHeaders.TENANT_ID)}
     )
     @Path("/echo")
-    @ActionBind(EchoAction.class)
+    @BindAction(EchoAction.class)
     String echo(String input, @Parameter(hidden = true) RequestContext context);
 
 }