+++ soffa-foundation-test-app/src/test/java/com/company/app/NoRabbitMQTest.java
@@ -1,7 +1,6 @@
 package com.company.app;
 
 import io.soffa.foundation.core.messages.AmqpClient;
-import io.soffa.foundation.core.messages.MessageDispatcher;
 import lombok.SneakyThrows;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
@@ -10,20 +9,16 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.ActiveProfiles;
 
 @ActiveProfiles("test")
-@SpringBootTest
+@SpringBootTest(properties = {"app.amqp.enabled=false"})
 public class NoRabbitMQTest {
 
     @Autowired(required = false)
     private AmqpClient amqpClient;
 
-    @Autowired(required = false)
-    private MessageDispatcher messageDispatcher;
-
     @SneakyThrows
     @Test
     public void testNoRabbitMQ() {
         Assertions.assertNull(amqpClient);
-        Assertions.assertNull(messageDispatcher);
     }
 
 }